# swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "маркет «Гофермарт»",
        "title": "Gophermart API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/api/user/balance": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Получение текущего баланса счёта баллов лояльности пользователя",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "balance"
                ],
                "summary": "Get user balance",
                "responses": {
                    "200": {
                        "description": "Баланс пользователя",
                        "schema": {
                            "$ref": "#/definitions/models.UserBalance"
                        }
                    },
                    "401": {
                        "description": "Пользователь не аутентифицирован",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/user/balance/withdraw": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Запрос на списание баллов с накопительного счёта в счёт оплаты нового заказа",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "withdrawals"
                ],
                "summary": "Withdraw balance",
                "parameters": [
                    {
                        "description": "Данные для списания",
                        "name": "withdraw",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.UserWithdrawals"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешная обработка запроса",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Неверный формат запроса",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Пользователь не аутентифицирован",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "402": {
                        "description": "На счету недостаточно средств",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "422": {
                        "description": "Неверный номер заказа",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/user/login": {
            "post": {
                "description": "Аутентификация пользователя в маркете «Гофермарт»",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "User login",
                "parameters": [
                    {
                        "description": "Данные для аутентификации",
                        "name": "credentials",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.UserCredentials"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Пользователь успешно аутентифицирован",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Неверный формат запроса",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Неверная пара логин/пароль",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/user/orders": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Получение списка загруженных пользователем номеров заказов, статусов их обработки и информации о начислениях.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "orders"
                ],
                "summary": "Get user orders",
                "responses": {
                    "200": {
                        "description": "Список заказов",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Order"
                            }
                        }
                    },
                    "204": {
                        "description": "Нет данных для ответа",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Пользователь не аутентифицирован",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Загрузка номера заказа для расчета баллов лояльности. Номер заказа проверяется алгоритмом Луна.",
                "consumes": [
                    "text/plain"
                ],
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "orders"
                ],
                "summary": "Upload order number",
                "parameters": [
                    {
                        "example": "\"12345678903\"",
                        "description": "Номер заказа",
                        "name": "order",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Заказ уже был загружен этим пользователем",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "202": {
                        "description": "Новый номер заказа принят в обработку",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Неверный формат запроса",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Пользователь не аутентифицирован",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "409": {
                        "description": "Номер заказа уже был загружен другим пользователем",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "422": {
                        "description": "Неверный формат номера заказа",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/user/register": {
            "post": {
                "description": "Регистрация нового пользователя в маркете «Гофермарт»",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "User registration",
                "parameters": [
                    {
                        "description": "Данные для регистрации",
                        "name": "credentials",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.UserCredentials"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Пользователь успешно зарегистрирован и аутентифицирован",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Неверный формат запроса",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "409": {
                        "description": "Логин уже занят",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/user/withdrawals": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Получение информации о выводе средств с накопительного счёта пользователем. Сортировка по времени вывода от новых к старым.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "withdrawals"
                ],
                "summary": "Get withdrawals history",
                "responses": {
                    "200": {
                        "description": "История выводов средств",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.UserWithdrawals"
                            }
                        }
                    },
                    "204": {
                        "description": "Нет ни одного списания",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Пользователь не аутентифицирован",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.Order": {
            "description": "Запрос на создание заказа",
            "type": "object",
            "properties": {
                "accrual": {
                    "type": "number",
                    "example": 100
                },
                "number": {
                    "type": "string",
                    "example": "2377225624"
                },
                "status": {
                    "type": "string",
                    "example": "NEW"
                },
                "uploaded_at": {
                    "type": "string",
                    "example": "2025-10-09T16:09:57+03:00"
                }
            }
        },
        "models.UserBalance": {
            "description": "Запрос на получения баланса",
            "type": "object",
            "properties": {
                "current": {
                    "type": "number",
                    "example": 1000
                },
                "withdrawn": {
                    "type": "number",
                    "example": 99
                }
            }
        },
        "models.UserCredentials": {
            "description": "Запрос на регистрацию и аутентификацию",
            "type": "object",
            "properties": {
                "login": {
                    "type": "string",
                    "example": "Djon"
                },
                "password": {
                    "type": "string",
                    "example": "123456"
                }
            }
        },
        "models.UserWithdrawals": {
            "description": "Запрос на историю выводов средств",
            "type": "object",
            "properties": {
                "order": {
                    "type": "string",
                    "example": "2377225624"
                },
                "processed_at": {
                    "type": "string",
                    "example": "2025-10-09T16:09:57+03:00"
                },
                "sum": {
                    "type": "number",
                    "example": 10
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "cookie"
        }
    },
    "tags": [
        {
            "name": "auth \"Аутентификация и регистрация\""
        },
        {
            "name": "orders \"Заказы\""
        },
        {
            "name": "balance \"Баланс\""
        },
        {
            "name": "withdrawals \"Списания\""
        }
    ]
}